+++ animation/src/Oval.java
@@ -1,7 +1,7 @@
 public class Oval extends AbstractShape {
   private final String type;
   
-  public Oval(String name, float pointX, float pointY, float X, float Y, int r, int g, int b) {
+  public Oval(String name, float pointX, float pointY, int X, int Y, int r, int g, int b) {
     super(name, pointX, pointY, X, Y,  r, g, b);
     this.type = "OVAL";
   }
