+++ animation/src/Transformation.java
@@ -1,98 +1,101 @@
 import java.util.List;
 
+/**
+ * Animation interface, represents methods implemented by the {@link AbstractTransformation} class.
+ */
 public interface Transformation {
   
   //GETTERS----------------------------------------------------------------------------------------
   
   /**
-   *
-   * @return
+   * Returns the {@link Shape}.
+   * @return a {@link Shape} object
    */
   Shape getShape();
   
   /**
-   *
-   * @return
+   * Returns the {@link Transformation}(s).
+   * @return a transformations
    */
   Transformation getTransformation();
   
   /**
-   *
-   * @return
+   * Returns the type of the {@link Transformation}.
+   * @return a string, type of transformation
    */
   String getTransformationType();
   
   /**
-   *
-   * @return
+   * Returns the time the transformation starts.
+   * @return int, time
    */
   int getTimeStart();
   
   /**
-   *
-   * @return
+   * Returns the time the transformation ends.
+   * @return int, time
    */
   int getTimeEnd();
   
   /**
-   *
-   * @return
+   * Returns the x-position/width the new shape should have.
+   * @return float, time
    */
   float getToX();
   
   /**
-   *
-   * @return
+   * Returns the y-position/length the new shape should have.
+   * @return float, time
    */
   float getToY();
   
   
   /**
-   *
-   * @return
+   * Returns the color the new shape should have.
+   * @return {@link Color}, color of new shape
    */
   Color getToColor();
   
   //OTHER------------------------------------------------------------------------------------------
   
   /**
-   *
-   * @return
+   * Returns a copy of a transformation object.
+   * @return a {@link Transformation} object
    */
   Transformation copy();
   
   /**
-   *
-   * @param toX
-   * @param toY
-   * @param timeStart
-   * @param timeEnd
-   * @return
+   * Returns a new {@link Shape} with provided x and y positions.
+   * @param toX x-position of the new shape object
+   * @param toY y-position of the new shape object
+   * @param timeStart time transformation starts
+   * @param timeEnd time transformation ends
+   * @return a new shape object
    */
   Shape moveShape(float toX, float toY, int timeStart, int timeEnd);
   
   /**
-   *
-   * @param toColor
-   * @param timeStart
-   * @param timeEnd
-   * @return
+   * Returns a new {@link Shape} with provided color.
+   * @param toColor {@link Color} of the new shape
+   * @param timeStart time transformation starts
+   * @param timeEnd time transformation ends
+   * @return a new shape object
    */
   Shape changeColor(Color toColor, int timeStart, int timeEnd);
   
   /**
-   *
-   * @param toX
-   * @param toY
-   * @param timeStart
-   * @param timeEnd
-   * @return
+   * Returns a new {@link Shape} with provided x and y positions.
+   * @param toX width of the new shape object
+   * @param toY length of the new shape object
+   * @param timeStart time transformation starts
+   * @param timeEnd time transformation ends
+   * @return a new shape object
    */
-  Shape scaleShape(float toX, float toY, int timeStart, int timeEnd);
+  Shape scaleShape(int toX, int toY, int timeStart, int timeEnd);
   
   /**
-   *
-   * @return
+   * Returns a string representation of the {@link Transformation} object.
+   * @return a string, decription of the transformation
    */
   String toString();
